Generalize frontend event table
This factors out most of the table code for re-use
with other assets, namely series and playlists.

Please note that the generalization of column sorting
and display will be done in later commits.
